Remove vpnotebook dependency
It is no longer needed (it used to import numpy, and a couple of other
things) because we handle imports in the core vpython code. Editorial:
custom kernels installed in --user can also mask kernels in --sys-prefix
(i.e. in a virtual environment) leading to confusion about which python
is actually running.