For demos that use ipywidgets use async version.
For demos that use ipywidgets use the async version of the vpython demo. This is because a vpython program can run forever in a cell and for many ipywidgets to work the cell must run to completion. Using async version of the demo vpython program will enable the ipywidgets to run properly.